[fix] alarm 생성을 object 클래스로 묶음
